:truck: Organize api calls
