+++ issuepoker-frontend/src/api/issue/delete-issue.ts
