+++ issuepoker-frontend/src/api/issue/get-filterOptions.ts
@@ -1,6 +1,6 @@
 import type { Filter } from "@/types/Filter.ts";
 
-import { defaultResponseHandler, getConfig } from "@/api/fetch-utils";
+import { defaultResponseHandler, getConfig } from "@/api/fetch-utils.ts";
 
 export function getFilterOptions(): Promise<Filter> {
   return fetch("api/backend-service/issues/filterOptions", getConfig())
