+++ issuepoker-frontend/src/api/misc/ad2image-avatar-client.ts
@@ -1,4 +1,4 @@
-import { AD2IMAGE_URL } from "@/constants";
+import { AD2IMAGE_URL } from "@/constants.ts";
 
 export class Ad2imageAvatarClient {
   base: string;
