+++ issuepoker-frontend/src/api/misc/health-client.ts
@@ -1,5 +1,5 @@
-import { defaultResponseHandler, getConfig } from "@/api/fetch-utils";
-import HealthState from "@/types/HealthState";
+import { defaultResponseHandler, getConfig } from "@/api/fetch-utils.ts";
+import HealthState from "@/types/HealthState.ts";
 
 export function checkGatewayHealth(): Promise<HealthState> {
   return fetch("actuator/health", getConfig())
