+++ issuepoker-frontend/src/api/vote/delete-vote-all.ts
