+++ issuepoker-frontend/src/api/vote/delete-vote.ts
