+++ issuepoker-frontend/src/components/IssueVoting.vue
@@ -140,12 +140,12 @@ import { isDefined } from "@vueuse/core";
 import { storeToRefs } from "pinia";
 import { onMounted, onUnmounted, ref, watch } from "vue";
 
-import { createVote } from "@/api/create-vote.ts";
-import { deleteAllVotes } from "@/api/delete-vote-all.ts";
-import { deleteVote } from "@/api/delete-vote.ts";
-import { subscribeVotes } from "@/api/fetch-votes.ts";
-import { setVoteResult } from "@/api/set-vote-result.ts";
-import { setVoteRevealed } from "@/api/set-vote-revealed.ts";
+import { createVote } from "@/api/vote/create-vote.ts";
+import { deleteAllVotes } from "@/api/vote/delete-vote-all.ts";
+import { deleteVote } from "@/api/vote/delete-vote.ts";
+import { fetchVotes } from "@/api/vote/subscribe-votes.ts";
+import { setVoteResult } from "@/api/vote/set-voteResult.ts";
+import { setVoteRevealed } from "@/api/vote/set-voteRevealed.ts";
 import YesNoDialog from "@/components/common/YesNoDialog.vue";
 import { STATUS_INDICATORS } from "@/constants.ts";
 import { useSnackbarStore } from "@/stores/snackbar.ts";
@@ -190,7 +190,7 @@ function fetchVotes() {
     return;
   }
   if (eventSource) eventSource.close();
-  eventSource = subscribeVotes(
+  eventSource = fetchVotes(
     props.issue,
     (content) => {
       votes.value = content;
